Create woodCompany.cpp
This will take in the users request for the wood type and dimensions then calculate the total.

//Current issue//
The running total is not calculating from the functions returning the boardPrice. Each boardPrice needs to be added together based on how much the users wants. Then when the user is done the input of the char T will end the program and output the total.